feat: inicio CRUD de transações
- Adiciona entidade transaction
- Realizar deposito
- Adiciona tabela transactions no BD
- Adiciona relação entre conta e transações

+++ src/main/java/bank/api/domain/account/DataGetAccount.java
@@ -2,7 +2,7 @@ package bank.api.domain.account;
 
 import java.math.BigDecimal;
 
-public record DataGetAccount(Long id, String numero, Double saldo) {
+public record DataGetAccount(Long id, String numero, BigDecimal saldo) {
     public DataGetAccount(Account account){
         this(account.getId(), account.getNumero(), account.getSaldo());
     }
